#11, removing many unnecessary classes those supported JUnit 3.
